Replace "unsigned int" by "Flags" in all AutoSegments collections.
* Change: In Anabatic::AutoSegments collections, change the type of all
    the flags that where in "unsigned int" (32 bits) to Flags (uint64_t)
    as there is now more than 32 flags for functions.
* New: In Ababatic::Constants, added new flag Flags::WithPerpands, which
    makes the number of flags tip over 32 bits, thus making mandatory
    to uses Flags and not unsigned int.
* New: In Anabatic::AutoSegments_Perpandiculars, manage a new flag
    Flags::WithDoglegs to allow to propagate through global segments that
    are connecteds via doglegs on local segments. Meaning that there is
    a good chance that they could be aligned.
      Slighly change the way we propagate on aligned segments: no longer
    check for VTee or HTee, but only for same direction and layer as
    master.
* New: In Anabatic & Katana, replace all the "int", "long" and their
    variants by the less implementation ambiguous "int32_t", "int64_t"
    (and variant). This should help to better detect bit trucation in
    flags.
      Use the type to give a hint about the flags kind:
        - Type "Flags", for flags shared among Anabatic & Katana
          functions/methods (may also appear in some objects states).
        - Type "uint32_t" for flags belonging to an object internal
          state of from Hurricane functions flags (those should be
          grouped in a Flag subclass in a perfect world).

+++ katana/src/Configuration.cpp
@@ -99,7 +99,7 @@ namespace Katana {
   { return new Configuration(*this); }
 
 
-  void  Configuration::setRipupLimit ( unsigned int type, unsigned int limit )
+  void  Configuration::setRipupLimit ( uint32_t type, uint32_t limit )
   {
     if ( type >= RipupLimitsTableSize ) {
       cerr << Error("setRipupLimit(): Bad ripup limit index: %ud (> %ud)."
@@ -111,7 +111,7 @@ namespace Katana {
   }
 
 
-  void  Configuration::setHTracksReservedLocal ( size_t reserved )
+  void  Configuration::setHTracksReservedLocal ( uint32_t reserved )
   {
     // if (reserved > getHEdgeCapacity())
     //   throw Error( "Configuration::setHTracksReservedLocal(): tracks reserved for local routing (%d) is greater than edge capacity %d."
@@ -121,7 +121,7 @@ namespace Katana {
   }
 
 
-  void  Configuration::setVTracksReservedLocal ( size_t reserved )
+  void  Configuration::setVTracksReservedLocal ( uint32_t reserved )
   {
     // if (reserved > 1.0)
     //   throw Error( "Configuration::setVTracksReservedLocal(): tracks reserved for local routing (%d) is greater than edge capacity %d."
@@ -131,7 +131,7 @@ namespace Katana {
   }
 
 
-  unsigned int  Configuration::getRipupLimit ( unsigned int type ) const
+  uint32_t  Configuration::getRipupLimit ( uint32_t type ) const
   {
     if ( type >= RipupLimitsTableSize ) {
       cerr << Error("getRipupLimit(): Bad ripup limit index: %u (> %u)."
